Navigation script
This is the principal node of a set of three nodes that allows a robot to wander in the Gazebo map recognizing objects and people. The navigation is very simple and naive, the agent goes forward and changes direction when it not possible to proceed, but it is a good way to test the efficiency of the detection system.